cleaned code a bit more

+++ 1_Example/main.cpp
@@ -5,7 +5,7 @@
 #include <iostream>
 
 struct Callback : public AudioCallback {
-    void prepare (uint64_t sampleRate) override {
+    void prepare (int sampleRate) override {
         for (auto& tremolo : tremolos)
             tremolo.prepareToPlay (static_cast<double> (sampleRate));
     }
